refactor: making getAutoComplete() static class function
